+++ frontend/app/src/components/settings/api-keys/ExchangeSettings.vue
@@ -33,6 +33,13 @@
         <template #item.location="{ item }">
           <location-display :identifier="item.location" />
         </template>
+        <template #item.syncEnabled="{ item }">
+          <v-switch
+            :input-value="!isNonSyncExchange(item)"
+            inset
+            @change="toggleSync(item)"
+          />
+        </template>
         <template #item.actions="{ item }">
           <row-actions
             :delete-tooltip="$t('exchange_settings.delete.tooltip')"
@@ -72,7 +79,7 @@
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
+import { Component, Mixins } from 'vue-property-decorator';
 import { DataTableHeader } from 'vuetify';
 import { mapActions, mapState } from 'vuex';
 import BaseExternalLink from '@/components/base/BaseExternalLink.vue';
@@ -83,7 +90,9 @@ import RowActions from '@/components/helper/RowActions.vue';
 import { exchangeName } from '@/components/history/consts';
 import RevealableInput from '@/components/inputs/RevealableInput.vue';
 import ExchangeKeysForm from '@/components/settings/api-keys/ExchangeKeysForm.vue';
+import SettingsMixin from '@/mixins/settings-mixin';
 import { ExchangePayload, ExchangeSetupPayload } from '@/store/balances/types';
+import { useNotifications } from '@/store/notifications';
 import { Nullable, Writeable } from '@/types';
 import { Exchange, SupportedExchange } from '@/types/exchanges';
 import { assert } from '@/utils/assertions';
@@ -117,7 +126,8 @@ const placeholder: () => ExchangePayload = () => ({
     ...mapActions('balances', ['setupExchange', 'removeExchange'])
   }
 })
-export default class ExchangeSettings extends Vue {
+export default class ExchangeSettings extends Mixins(SettingsMixin) {
+  nonSyncingExchanges: Exchange[] = [];
   pendingRemoval: Nullable<Exchange> = null;
   confirmation: boolean = false;
   setupExchange!: (payload: ExchangeSetupPayload) => Promise<boolean>;
@@ -131,11 +141,62 @@ export default class ExchangeSettings extends Vue {
   pending: boolean = false;
   binancePairs: string[] = [];
 
+  created() {
+    this.nonSyncingExchanges = this.generalSettings.nonSyncingExchanges;
+  }
+
   mounted() {
     const { currentRoute } = this.$router;
     if (currentRoute.query.add) {
       this.addExchange();
     }
+  }
+
+  findNonSyncExchangeIndex(exchange: Exchange) {
+    return this.nonSyncingExchanges.findIndex((item: Exchange) => {
+      return item.name === exchange.name && item.location === exchange.location;
+    });
+  }
+
+  isNonSyncExchange(exchange: Exchange) {
+    return this.findNonSyncExchangeIndex(exchange) > -1;
+  }
+
+  async toggleSync(exchange: Exchange) {
+    const index = this.findNonSyncExchangeIndex(exchange);
+
+    let data = [...this.nonSyncingExchanges];
+
+    let enable = true;
+
+    if (index > -1) {
+      enable = false;
+      data.splice(index);
+    } else {
+      data.push({ location: exchange.location, name: exchange.name });
+    }
+
+    const status = await this.settingsUpdate({
+      nonSyncingExchanges: data
+    });
+
+    if (!status.success) {
+      const { notify } = useNotifications();
+      notify({
+        title: this.$t('exchange_settings.sync.messages.title').toString(),
+        message: this.$t('exchange_settings.sync.messages.description', {
+          action: enable
+            ? this.$t('exchange_settings.sync.messages.enable')
+            : this.$t('exchange_settings.sync.messages.disable'),
+          location: exchange.location,
+          name: exchange.name,
+          message: status.message
+        }).toString(),
+        display: true
+      });
+    }
+
+    this.nonSyncingExchanges = this.generalSettings.nonSyncingExchanges;
   }
 
   get dialogTitle(): string {
@@ -167,6 +228,10 @@ export default class ExchangeSettings extends Vue {
     {
       text: this.$t('exchange_settings.header.name').toString(),
       value: 'name'
+    },
+    {
+      text: this.$t('exchange_settings.header.sync_enabled').toString(),
+      value: 'syncEnabled'
     },
     {
       text: this.$t('exchange_settings.header.actions').toString(),